+++ scaffold/generate/resource/resource.rb
