+++ scaffold/generate/resource/resource/create.rb
