+++ scaffold/generate/resource/resource/create/authorizer.rb
