+++ scaffold/generate/resource/resource/create/control.rb
