+++ scaffold/generate/resource/resource/create/denormalizer.rb
