+++ scaffold/generate/resource/resource/create/normalizer.rb
