+++ scaffold/generate/resource/resource/destroy.rb
