+++ scaffold/generate/resource/resource/destroy/authorizer.rb
