+++ scaffold/generate/resource/resource/destroy/control.rb
