+++ scaffold/generate/resource/resource/destroy/denormalizer.rb
