+++ scaffold/generate/resource/resource/destroy/normalizer.rb
