+++ scaffold/generate/resource/resource/destroy/validator.rb
