+++ scaffold/generate/resource/resource/endpoint.rb
