+++ scaffold/generate/resource/resource/list.rb
