+++ scaffold/generate/resource/resource/list/authorizer.rb
