+++ scaffold/generate/resource/resource/list/control.rb
