+++ scaffold/generate/resource/resource/list/denormalizer.rb
