+++ scaffold/generate/resource/resource/list/normalizer.rb
