+++ scaffold/generate/resource/resource/model.rb
