+++ scaffold/generate/resource/resource/null.rb
