+++ scaffold/generate/resource/resource/presenter.rb
