+++ scaffold/generate/resource/resource/show.rb
