+++ scaffold/generate/resource/resource/show/authorizer.rb
