+++ scaffold/generate/resource/resource/show/control.rb
