+++ scaffold/generate/resource/resource/show/denormalizer.rb
