+++ scaffold/generate/resource/resource/show/normalizer.rb
