+++ scaffold/generate/resource/resource/show/validator.rb
