+++ scaffold/generate/resource/resource/update.rb
