+++ scaffold/generate/resource/resource/update/control.rb
