+++ scaffold/generate/resource/resource/update/denormalizer.rb
