+++ scaffold/generate/resource/resource/update/normalizer.rb
